TBI participant generator
This code creates 100 participants and assigns them a numerical ID from 1 to 100. Then, each participant is randomly given an age between 20 and 100 years. Each person is randomly given a gender of either male or female. Then, each person is randomly assigned a TBI severity of either mild, moderate, or severe. Once each of these things is generated, the variables for Suicidal Ideation and Depression are randomly generated but weighted in such a way that those with more severe TBI, those who are male, and those who are older, will have a higher average suicidal ideation score.